Update quiz_program.py
Looks much better! More organization in the data, and in the code.    ID values are very often a good thing to add to data in your program, in the same way that primary key fields are almost always a good thing to add to a database table.  

Some notes and feedback.  This isn't a working version - I didn't edit the space category, only the art category.  You don't need to merge this, it's just easier to add notes on your code this way.  

The data structure is making things more difficult when you access data, and there's possible bugs that might happen - what if the IDs are not sequential, or they don't start at 1?  Converting strings to ints is also a possible source of bugs.   Would there be a better way to structure this, that makes it easier to access the data?  

Suggested edits to the data structure - questions key does relate to answers value, but this makes more work when you read the question. If a dictionary only has one key-value pair, restructuring or using a different data structure may be better. 

Think about names - make sure they accurately reflect what your function does. 

Functions should have one specific task - processing data and printing are two different things. 

Code shouldn't be in a loop if it will always happen exactly one time. 

Please let me know if you have any questions on any of this. Thank you!